+++ src/main/java/org/example/company/tcs/techcellshop/dto/response/DeviceSummaryResponse.java
@@ -1,4 +1,6 @@
 package org.example.company.tcs.techcellshop.dto.response;
 
-public record DeviceSummaryResponse(Long idDevice, String nameDevice, Double devicePrice) {
+import java.math.BigDecimal;
+
+public record DeviceSummaryResponse(Long idDevice, String nameDevice, BigDecimal devicePrice) {
 }
